refactor: split naver map helper responsibilities
