+++ src/components/naver-map/mapSdk.ts
@@ -1,0 +1,34 @@
+declare global {
+  interface Window {
+    naver?: any;
+  }
+}
+
+let naverScriptPromise: Promise<any> | null = null;
+
+export function loadNaverMaps(clientId: string) {
+  if (window.naver?.maps) {
+    return Promise.resolve(window.naver.maps);
+  }
+
+  if (naverScriptPromise) {
+    return naverScriptPromise;
+  }
+
+  naverScriptPromise = new Promise((resolve, reject) => {
+    const script = document.createElement('script');
+    script.src = `https://oapi.map.naver.com/openapi/v3/maps.js?ncpKeyId=${clientId}`;
+    script.async = true;
+    script.onload = () => {
+      if (window.naver?.maps) {
+        resolve(window.naver.maps);
+        return;
+      }
+      reject(new Error('네이버 지도 SDK를 찾지 못했어요.'));
+    };
+    script.onerror = () => reject(new Error('네이버 지도 SDK를 불러오지 못했어요.'));
+    document.head.appendChild(script);
+  });
+
+  return naverScriptPromise;
+}
